feat: add NotFound page, refactor dashboard layout and agent management components, implement agents filtering and pagination, and integrate TRPC for data handling
