+++ app/(dashboard)/agents/page.tsx
@@ -1,11 +1,11 @@
-import { auth } from "@/lib/auth";
-import { loadSearchParams } from "@/modules/agents/params";
-import AgentsListHeader from "@/modules/agents/ui/components/agents-list-header";
+import { loadSearchParams } from "@/app/(dashboard)/agents/lib/params";
+import AgentsListHeader from "@/app/(dashboard)/agents/ui/components/agents-list-header";
 import {
 	AgentView,
 	AgentViewError,
 	AgentViewLoader,
-} from "@/modules/agents/ui/views/agent-view";
+} from "@/app/(dashboard)/agents/ui/views/agent-view";
+import { auth } from "@/lib/auth";
 import { getQueryClient, trpc } from "@/trpc/server";
 import { HydrationBoundary, dehydrate } from "@tanstack/react-query";
 import { headers } from "next/headers";
