+++ app/(dashboard)/page.tsx
@@ -1,7 +1,7 @@
-import Home from "@/components/home/home";
 import { auth } from "@/lib/auth";
 import { headers } from "next/headers";
 import { redirect } from "next/navigation";
+import { Home } from "./ui/views/home";
 
 export default async function HomePage() {
 	const session = await auth.api.getSession({
