+++ app/(dashboard)/ui/views/home.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-export default function Home() {
+export function Home() {
 	return (
 		<main className="flex flex-col p-4 gap-y-4">
 			<p>Home View</p>
