+++ trpc/routers/_app.ts
@@ -1,4 +1,4 @@
-import { agentsRouter } from "@/modules/agents/server/procedure";
+import { agentsRouter } from "@/app/(dashboard)/agents/server/procedure";
 import { createTRPCRouter } from "../init";
 
 export const appRouter = createTRPCRouter({
